VXI11_CLIENTS is passed to open/close functons
VXI11_CLIENTS is no longer a global variable but is instead passed to
the open/close functions. In a multithreaded application the global
variable causes problems. Instead it is up to the calling code to
protect this variable.
